# personal-budget-1/docs/api-docs.json
{
    "openapi": "3.0.1",
    "info": {
        "description": "Personal Budget 1 Project (envelope budgeting). Codecademy Node.js Career Path.",
        "version": "1.0.0",
        "title": "Personal Budget 1",
        "contact": {
            "email": "yehor.zemzyulin@gmail.com"
        },
        "license": {}
    },
    "servers": [{
        "url": "/api"
    }],
    "paths" : {
        "/envelopes" : {
            "get" : {
                "description": "Returns all envelopes",
                "summary": "Retrieve all envelopes",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "examples": {
                                    "envelopes": {
                                        "value": [{"id":"1","category":"Investing","budget":3020},
                                            {"id":"2","category":"Housing","budget":2250},
                                            {"id":"3","category":"Personal Spending","budget":2510}]
                                    }
                                }
                            }
                        }
                    }
                },
                "tags": ["Get envelopes"]
            },
            "post": {
                "description": "Creates a new envelope",
                "summary": "Add envelope",
                "requestBody": {
                    "description": "Budget category and amount for new envelope",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/envelope"
                            },
                            "example": {"category":"Investing","budget":3020}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Returns created envelope"
                    },
                    "400": {
                        "description": "Invalid input"
                    }
                },
                "tags": ["Add envelope"]
            }
        },
        "/envelopes/{id}": {
            "get": {
                "description": "Returns envelope based on ID",
                "summary": "Find envelope by ID",
                "parameters": [{
                    "name": "id",
                    "in": "path",
                    "description": "ID of envelope",
                    "required": true,
                    "schema": {
                        "type": "integer"
                    },
                    "example": 1
                }],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "category": {
                                            "description": "Budget category",
                                            "type": "string"
                                        },
                                        "budget": {
                                            "description": "Amount to be allocated to envelope",
                                            "type": "integer"
                                        },
                                        "id": {
                                            "description": "ID of envelope",
                                            "type": "integer"
                                        }
                                    }
                                },
                                "example": {"category":"Investing","budget":3020,"id":"1"}
                            }
                        }
                    },
                    "404": {
                        "description": "ID not found"
                    }
                },
                "tags": ["Get envelopes"]
            },
            "put": {
                "description": "Updates envelope based on ID",
                "summary": "Update envelope by ID",
                "parameters": [{
                    "name": "id",
                    "in": "path",
                    "description": "ID of envelope",
                    "required": true,
                    "schema": {
                        "type": "integer"
                    },
                    "example": 1
                }],
                "requestBody": {
                    "description": "Budget category and amount to update envelope",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/envelope"
                            },
                            "example": {"category":"Investing","budget":3020}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "category": {
                                            "description": "Budget category",
                                            "type": "string"
                                        },
                                        "budget": {
                                            "description": "Amount to be allocated to envelope",
                                            "type": "integer"
                                        },
                                        "id": {
                                            "description": "ID of envelope",
                                            "type": "integer"
                                        }
                                    }
                                },
                                "example": {"category":"Investing","budget":3020,"id":"1"}
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input"
                    },
                    "404": {
                        "description": "ID not found"
                    }
                },
                "tags": ["Update envelope"]
            },
            "delete": {
                "description": "Deletes envelope based on ID",
                "summary": "Delete envelope by ID",
                "parameters": [{
                    "name": "id",
                    "in": "path",
                    "description": "ID of envelope",
                    "required": true,
                    "schema": {
                        "type": "integer"
                    },
                    "example": 1
                }],
                "responses": {
                    "204": {
                        "description": "Envelope deleted"
                    },
                    "404": {
                        "description": "ID not found"
                    }
                },
                "tags": ["Delete envelope"]
            }
        },
        "envelopes/transfer/{fromId}/{toId}": {
            "post": {
                "description": "Transfer funds between envelopes",
                "summary": "Transfer funds between envelopes by IDs",
                "parameters": [
                    {
                        "name": "fromId",
                        "in": "path",
                        "description": "ID of envelope from which to withdraw",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    },
                    {
                        "name": "toId",
                        "in": "path",
                        "description": "ID of envelope to which to deposit",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 2
                    }
                ],
                "requestBody": {
                    "description": "Amount to transfer",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                    "properties": {
                                        "amount": {
                                            "description": "Amount to transfer between envelopes",
                                            "type": "integer"
                                        }
                                    }
                            },
                            "example": {"amount":3020}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Transfer completed"
                    },
                    "400": {
                        "description": "Invalid input"
                    }
                },
                "tags": ["Transfer funds"]
            }
        }
    },
    "components": {
        "envelope": {
            "type": "object",
            "properties": {
                "category": {
                    "description": "Budget category",
                    "type": "string"
                },
                "budget": {
                    "description": "Amount to be allocated to envelope",
                    "type": "integer"
                }
            }
        }
    }
}